requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
distro==1.9.0
Flask==3.1.0
flask-cors==6.0.1
Flask-Session==0.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.52.0
google-generativeai==0.8.5
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
msgspec==0.19.0
openai==1.60.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
spotipy==2.24.0
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
